camera separation frametime fix
camera separation frametime fix

+++ ctr-tools/ctrviewer/Engine/EngineSettings.cs
@@ -74,7 +74,7 @@ namespace ctrviewer.Engine
 
         public bool StereoPair { get; set; } = false;
         public string PlayerModel { get; set; } = "crash";
-        public int StereoPairSeparation { get; set; } = 20;
+        public float StereoPairSeparation { get; set; } = 20;
         public bool ShowCamPos { get; set; } = false;
         public bool UseLowLod { get; set; } = false;
         public bool ShowConsole { get; set; } = false;
